requirements.txt
scrapy>=2.13
scikit-learn>=1.6
flask>=3.1
nltk>=3.9
beautifulsoup4>=4.12
lxml>=5.2
pandas>=2.2
faiss-cpu>=1.8.0
gensim>=4.3
python-dotenv>=1.0
pyyaml>=6.0.2
